queue/worker: Move the .enqueue() method to the Queue module
Change-type: patch
