+++ lib/queue/errors.js
@@ -18,7 +18,10 @@ const _ = require('lodash')
 const typedErrors = require('typed-errors')
 
 _.each([
-	'QueueNoRequest'
+	'QueueNoRequest',
+	'QueueInvalidRequest',
+	'QueueInvalidAction',
+	'QueueInvalidSession'
 ], (error) => {
 	exports[error] = typedErrors.makeTypedError(error)
 })
