+++ lib/queue/index.js
@@ -14,7 +14,10 @@
  * limitations under the License.
  */
 
+const Bluebird = require('bluebird')
+const uuid = require('uuid/v4')
 const logger = require('../logger').getLogger(__filename)
+const jellyscript = require('../jellyscript')
 const events = require('./events')
 const errors = require('./errors')
 const MemoryQueue = require('./memory-queue')
@@ -34,6 +37,7 @@ module.exports = class Queue {
 		this.jellyfish = jellyfish
 		this.session = session
 		this.data = new MemoryQueue()
+		this.errors = errors
 	}
 
 	/**
@@ -142,6 +146,7 @@ module.exports = class Queue {
 	 * @function
 	 * @public
 	 *
+	 * @param {String} session - session
 	 * @param {Object} options - options
 	 * @param {String} options.action - action slug
 	 * @param {String} options.card - action input card id
@@ -150,8 +155,75 @@ module.exports = class Queue {
 	 * @param {Date} [options.currentDate] - current date
 	 * @param {String} [options.originator] - card id that originated this action
 	 * @param {Object} [options.context] - execution context
+	 * @returns {Object} action request
 	 */
-	async enqueue (options) {
-		await this.data.enqueue(options)
+	async enqueue (session, options) {
+		logger.info(options.context, 'Enqueueing request', {
+			length: await this.length(),
+			request: {
+				action: options.action,
+				card: options.card,
+				type: options.type
+			}
+		})
+
+		if (!options.type) {
+			throw new errors.QueueInvalidRequest('Missing input card type')
+		}
+
+		const cards = await Bluebird.props({
+			target: this.jellyfish.getCardBySlug(
+				options.context, this.session, options.card, {
+					type: options.type
+				}),
+			action: this.jellyfish.getCardBySlug(
+				options.context, this.session, options.action, {
+					type: 'action'
+				}),
+			session: this.jellyfish.getCardById(
+				options.context, session, session, {
+					type: 'session'
+				})
+		})
+
+		if (!cards.session) {
+			throw new errors.QueueInvalidSession(
+				`No such session: ${session}`, options.context)
+		}
+
+		if (!cards.action) {
+			throw new errors.QueueInvalidAction(
+				`No such action: ${options.action}`, options.context)
+		}
+
+		const targetId = cards.target ? cards.target.id : options.card
+		const date = options.currentDate || new Date()
+		const request = {
+			id: uuid(),
+			context: options.context,
+			card: targetId,
+			type: options.type,
+			actor: cards.session.data.actor,
+			action: cards.action,
+			originator: options.originator,
+			timestamp: date.toISOString(),
+			epoch: date.valueOf(),
+			arguments: jellyscript.evaluateObject({
+				type: 'object',
+				properties: cards.action.data.arguments,
+				additionalProperties: false,
+				required: cards.action.data.required ||
+					Object.keys(cards.action.data.arguments)
+			}, options.arguments)
+		}
+
+		await this.data.enqueue(request)
+		return {
+			id: request.id,
+			action: cards.action.id,
+			card: targetId,
+			type: options.type,
+			actor: cards.session.data.actor
+		}
 	}
 }
