+++ lib/server/routes.js
@@ -221,7 +221,7 @@ exports.bindRoutes = (jellyfish, worker, queue, app, socketServer) => {
 				throw new Error(`No type card: ${EXTERNAL_EVENT_TYPE}`)
 			}
 
-			return worker.enqueue(jellyfish.sessions.admin, {
+			return queue.enqueue(jellyfish.sessions.admin, {
 				action: 'action-create-card',
 				card: typeCard.id,
 				type: typeCard.type,
@@ -310,7 +310,7 @@ exports.bindRoutes = (jellyfish, worker, queue, app, socketServer) => {
 		}
 
 		action.context = ctx
-		return worker.enqueue(request.sessionToken, action).then((id) => {
+		return queue.enqueue(request.sessionToken, action).then((id) => {
 			return queue.waitResults(ctx, id)
 		}).then(async (results) => {
 			logger.info(ctx, 'Got action results', results)
@@ -386,7 +386,7 @@ exports.bindRoutes = (jellyfish, worker, queue, app, socketServer) => {
 		const currentDate = new Date()
 		const session = jellyfish.sessions.admin
 
-		return worker.enqueue(session, {
+		return queue.enqueue(session, {
 			action: 'action-create-card',
 			card: 'ping',
 			type: 'type',
