+++ lib/worker/index.js
@@ -17,12 +17,10 @@
 const Bluebird = require('bluebird')
 const errio = require('errio')
 const _ = require('lodash')
-const uuid = require('uuid/v4')
 const errors = require('./errors')
 const executor = require('./executor')
 const utils = require('./utils')
 const triggers = require('./triggers')
-const jellyscript = require('../jellyscript')
 const logger = require('../logger').getLogger(__filename)
 
 module.exports = class Worker {
@@ -123,7 +121,7 @@ module.exports = class Worker {
 			attachEvents: options.attachEvents,
 			setTriggers: this.setTriggers.bind(this),
 			executeAction: async (executeSession, actionRequest) => {
-				return this.enqueue(executeSession, actionRequest)
+				return this.queue.enqueue(executeSession, actionRequest)
 			}
 		}, card)
 	}
@@ -216,101 +214,6 @@ module.exports = class Worker {
 	 */
 	getTriggers () {
 		return this.triggers
-	}
-
-	/**
-	 * @summary Enqueue an action request
-	 * @function
-	 * @public
-	 *
-	 * @param {String} session - session id
-	 * @param {Object} options - options
-	 * @param {String} options.action - action slug
-	 * @param {String} options.card - action input card id
-	 * @param {String} options.type - action input card type
-	 * @param {Object} options.arguments - action arguments
-	 * @param {Date} [options.currentDate] - current date
-	 * @param {String} [options.originator] - card id that originated this action
-	 * @param {Object} [options.context] - execution context
-	 * @returns {Object} action request
-	 *
-	 * @example
-	 * const worker = new Worker({ ... })
-	 * const request = await worker.enqueue('4a962ad9-20b5-4dd8-a707-bf819593cc84', {
-	 *   action: 'action-create-card',
-	 *   type: 'card',
-	 *   card: '033d9184-70b2-4ec9-bc39-9a249b186422',
-	 *   originator: '2ebb5e46-ef8a-485b-ba04-e8266d8c0677',
-	 *   arguments: {
-	 *     properties: {
-	 *       data: {
-	 *         foo: 'bar'
-	 *       }
-	 *     }
-	 *   }
-	 * })
-	 */
-	async enqueue (session, options) {
-		logger.info(options.context, 'Enqueueing request', {
-			request: {
-				action: options.action,
-				card: options.card,
-				type: options.type
-			},
-			length: await this.queue.length()
-		})
-
-		if (!options.type) {
-			throw new errors.WorkerInvalidActionRequest('Missing input card type')
-		}
-
-		const cards = await Bluebird.props({
-			target: this.jellyfish.getCardBySlug(options.context, session, options.card, {
-				type: options.type
-			}),
-			action: this.jellyfish.getCardBySlug(options.context, session, options.action, {
-				type: 'action'
-			}),
-			session: this.jellyfish.getCardById(options.context, session, session, {
-				type: 'session'
-			})
-		})
-
-		if (!cards.session) {
-			throw new errors.WorkerNoElement(`No such session: ${session}`, options.context)
-		}
-
-		if (!cards.action) {
-			throw new errors.WorkerInvalidAction(`No such action: ${options.action}`, options.context)
-		}
-
-		const targetId = cards.target ? cards.target.id : options.card
-
-		const date = options.currentDate || new Date()
-		const request = {
-			id: uuid(),
-			context: options.context,
-			card: targetId,
-			type: options.type,
-			actor: cards.session.data.actor,
-			action: cards.action,
-			originator: options.originator,
-			timestamp: date.toISOString(),
-			epoch: date.valueOf(),
-			arguments: jellyscript.evaluateObject(
-				utils.getActionArgumentsSchema(cards.action),
-				options.arguments
-			)
-		}
-
-		await this.queue.enqueue(request)
-		return {
-			id: request.id,
-			action: cards.action.id,
-			card: targetId,
-			type: options.type,
-			actor: cards.session.data.actor
-		}
 	}
 
 	/**
@@ -477,7 +380,7 @@ module.exports = class Worker {
 				return null
 			}
 
-			return this.enqueue(session, request)
+			return this.queue.enqueue(session, request)
 		}, {
 			concurrency: 5
 		})
