+++ test/integration/sync/helpers.js
@@ -133,13 +133,14 @@ const webhookScenario = async (test, testCase, integration, stub) => {
 				}
 			})
 
-		const request = await test.context.worker.enqueue(test.context.session, {
-			context: test.context.context,
-			action: 'action-integration-import-event',
-			card: event.id,
-			type: event.type,
-			arguments: {}
-		})
+		const request = await test.context.queue.enqueue(
+			test.context.session, {
+				context: test.context.context,
+				action: 'action-integration-import-event',
+				card: event.id,
+				type: event.type,
+				arguments: {}
+			})
 
 		await test.context.flush(test.context.session)
 		const result = await test.context.queue.waitResults(
