+++ test/unit/worker/index.spec.js
@@ -25,203 +25,10 @@ ava.beforeEach(async (test) => {
 
 ava.afterEach(helpers.worker.afterEach)
 
-ava('.enqueue() should increment the queue length by one', async (test) => {
-	const typeCard = await test.context.jellyfish.getCardBySlug(test.context.context, test.context.session, 'card')
-	await test.context.worker.enqueue(test.context.session, {
-		action: 'action-create-card',
-		context: test.context.context,
-		card: typeCard.id,
-		type: typeCard.type,
-		arguments: {
-			properties: {
-				version: '1.0.0',
-				data: {
-					foo: 'bar'
-				}
-			}
-		}
-	})
-
-	const length = await test.context.queue.length()
-	test.is(length, 1)
-})
-
-ava('.enqueue() should include the actor from the passed session', async (test) => {
-	const typeCard = await test.context.jellyfish.getCardBySlug(test.context.context, test.context.session, 'card')
-	const session = await test.context.jellyfish.getCardById(test.context.context, test.context.session, test.context.session)
-	await test.context.worker.enqueue(test.context.session, {
-		action: 'action-create-card',
-		context: test.context.context,
-		card: typeCard.id,
-		type: typeCard.type,
-		arguments: {
-			properties: {
-				version: '1.0.0',
-				slug: 'foo',
-				data: {
-					foo: 'bar'
-				}
-			}
-		}
-	})
-
-	const request = await test.context.queue.dequeue()
-	test.is(session.data.actor, request.actor)
-})
-
-ava('.enqueue() should include the whole passed action', async (test) => {
-	const typeCard = await test.context.jellyfish.getCardBySlug(
-		test.context.context, test.context.session, 'card')
-	const actionCard = await test.context.jellyfish.getCardBySlug(
-		test.context.context, test.context.session, 'action-create-card')
-	await test.context.worker.enqueue(test.context.session, {
-		action: 'action-create-card',
-		context: test.context.context,
-		card: typeCard.id,
-		type: typeCard.type,
-		arguments: {
-			properties: {
-				version: '1.0.0',
-				slug: 'foo',
-				data: {
-					foo: 'bar'
-				}
-			}
-		}
-	})
-
-	const request = await test.context.queue.dequeue()
-	test.deepEqual(request.action, actionCard)
-})
-
-ava('.enqueue() should set an originator', async (test) => {
-	const typeCard = await test.context.jellyfish.getCardBySlug(test.context.context, test.context.session, 'card')
-	await test.context.worker.enqueue(test.context.session, {
-		action: 'action-create-card',
-		context: test.context.context,
-		card: typeCard.id,
-		type: typeCard.type,
-		originator: '4a962ad9-20b5-4dd8-a707-bf819593cc84',
-		arguments: {
-			properties: {
-				slug: 'foo',
-				version: '1.0.0'
-			}
-		}
-	})
-
-	const request = await test.context.queue.dequeue()
-	test.is(request.originator, '4a962ad9-20b5-4dd8-a707-bf819593cc84')
-})
-
-ava('.enqueue() should take a current date', async (test) => {
-	const typeCard = await test.context.jellyfish.getCardBySlug(test.context.context, test.context.session, 'card')
-	const date = new Date()
-
-	await test.context.worker.enqueue(test.context.session, {
-		action: 'action-create-card',
-		context: test.context.context,
-		card: typeCard.id,
-		type: typeCard.type,
-		currentDate: date,
-		arguments: {
-			properties: {
-				slug: 'foo',
-				version: '1.0.0'
-			}
-		}
-	})
-
-	const request = await test.context.queue.dequeue()
-	test.is(request.timestamp, date.toISOString())
-})
-
-ava('.enqueue() should set a present timestamp', async (test) => {
-	const currentDate = new Date()
-	const typeCard = await test.context.jellyfish.getCardBySlug(test.context.context, test.context.session, 'card')
-	await test.context.worker.enqueue(test.context.session, {
-		action: 'action-create-card',
-		context: test.context.context,
-		card: typeCard.id,
-		type: typeCard.type,
-		arguments: {
-			properties: {
-				version: '1.0.0',
-				slug: 'foo',
-				data: {
-					foo: 'bar'
-				}
-			}
-		}
-	})
-
-	const request = await test.context.queue.dequeue()
-	test.true(new Date(request.timestamp) >= currentDate)
-})
-
-ava('.enqueue() should throw if the action was not found', async (test) => {
-	const typeCard = await test.context.jellyfish.getCardBySlug(test.context.context, test.context.session, 'card')
-	await test.throwsAsync(test.context.worker.enqueue(test.context.session, {
-		action: 'action-foo-bar',
-		context: test.context.context,
-		card: typeCard.id,
-		type: typeCard.type,
-		arguments: {
-			properties: {
-				version: '1.0.0',
-				slug: 'foo',
-				data: {
-					foo: 'bar'
-				}
-			}
-		}
-	}), test.context.worker.errors.WorkerInvalidAction)
-})
-
-ava('.enqueue() should throw if the session was not found', async (test) => {
-	const typeCard = await test.context.jellyfish.getCardBySlug(test.context.context, test.context.session, 'card')
-	await test.throwsAsync(test.context.worker.enqueue('aaaaaaaa-aaaa-aaaa-aaaa-aaaaaaaaaaaa', {
-		action: 'action-create-card',
-		context: test.context.context,
-		card: typeCard.id,
-		type: typeCard.type,
-		arguments: {
-			properties: {
-				version: '1.0.0',
-				slug: 'foo',
-				data: {
-					foo: 'bar'
-				}
-			}
-		}
-	}), test.context.jellyfish.errors.JellyfishNoElement)
-})
-
-ava('.enqueue() should not store the password in the queue when using action-create-user', async (test) => {
-	const typeCard = await test.context.jellyfish.getCardBySlug(test.context.context, test.context.session, 'user')
-	await test.context.worker.enqueue(test.context.session, {
-		action: 'action-create-user',
-		context: test.context.context,
-		card: typeCard.id,
-		type: typeCard.type,
-		arguments: {
-			email: 'johndoe@example.com',
-			username: 'user-johndoe',
-			hash: {
-				string: 'foobarbaz',
-				salt: 'user-johndoe'
-			}
-		}
-	})
-
-	const request = await test.context.queue.dequeue()
-	test.falsy(request.arguments.hash.string)
-	test.falsy(request.arguments.hash.salt)
-})
-
-ava('.enqueue() should not store the password in the queue when using action-create-session', async (test) => {
-	const userCard = await test.context.jellyfish.getCardBySlug(test.context.context, test.context.session, 'user')
-	const createUserRequest = await test.context.worker.enqueue(test.context.session, {
+ava('should not store the password in the queue when using action-create-session', async (test) => {
+	const userCard = await test.context.jellyfish.getCardBySlug(
+		test.context.context, test.context.session, 'user')
+	const createUserRequest = await test.context.queue.enqueue(test.context.session, {
 		action: 'action-create-user',
 		context: test.context.context,
 		card: userCard.id,
@@ -241,7 +48,7 @@ ava('.enqueue() should not store the password in the queue when using action-cre
 		test.context.context, createUserRequest)
 	test.false(result.error)
 
-	await test.context.worker.enqueue(test.context.session, {
+	await test.context.queue.enqueue(test.context.session, {
 		action: 'action-create-session',
 		context: test.context.context,
 		card: result.data.id,
@@ -261,11 +68,13 @@ ava('.enqueue() should not store the password in the queue when using action-cre
 	test.falsy(request.arguments.password.hash.salt)
 })
 
-ava('enqueue() should fail to create an event with an action-create-card', async (test) => {
-	const cardType = await test.context.jellyfish.getCardBySlug(test.context.context, test.context.session, 'card')
-	const typeType = await test.context.jellyfish.getCardBySlug(test.context.context, test.context.session, 'type')
+ava('should fail to create an event with an action-create-card', async (test) => {
+	const cardType = await test.context.jellyfish.getCardBySlug(
+		test.context.context, test.context.session, 'card')
+	const typeType = await test.context.jellyfish.getCardBySlug(
+		test.context.context, test.context.session, 'type')
 
-	const id = await test.context.worker.enqueue(test.context.session, {
+	const id = await test.context.queue.enqueue(test.context.session, {
 		action: 'action-create-card',
 		context: test.context.context,
 		card: typeType.id,
@@ -306,7 +115,7 @@ ava('enqueue() should fail to create an event with an action-create-card', async
 		test.context.context, id)
 	test.false(typeResult.error)
 
-	const threadId = await test.context.worker.enqueue(test.context.session, {
+	const threadId = await test.context.queue.enqueue(test.context.session, {
 		action: 'action-create-card',
 		context: test.context.context,
 		card: typeResult.data.id,
@@ -327,7 +136,7 @@ ava('enqueue() should fail to create an event with an action-create-card', async
 		test.context.context, threadId)
 	test.false(threadResult.error)
 
-	await test.context.worker.enqueue(test.context.session, {
+	await test.context.queue.enqueue(test.context.session, {
 		action: 'action-create-card',
 		card: cardType.id,
 		context: test.context.context,
@@ -354,32 +163,9 @@ ava('enqueue() should fail to create an event with an action-create-card', async
 	)
 })
 
-ava('.dequeue() should reduce the queue length', async (test) => {
-	const typeCard = await test.context.jellyfish.getCardBySlug(test.context.context, test.context.session, 'card')
-	await test.context.worker.enqueue(test.context.session, {
-		action: 'action-create-card',
-		context: test.context.context,
-		card: typeCard.id,
-		type: typeCard.type,
-		arguments: {
-			properties: {
-				slug: 'foo',
-				version: '1.0.0',
-				data: {
-					foo: 1
-				}
-			}
-		}
-	})
-
-	await test.context.queue.dequeue()
-	const length = await test.context.queue.length()
-	test.is(length, 0)
-})
-
 ava('.execute() should execute an action', async (test) => {
 	const typeCard = await test.context.jellyfish.getCardBySlug(test.context.context, test.context.session, 'card')
-	const request = await test.context.worker.enqueue(test.context.session, {
+	const request = await test.context.queue.enqueue(test.context.session, {
 		action: 'action-create-card',
 		context: test.context.context,
 		card: typeCard.id,
@@ -409,7 +195,7 @@ ava('.execute() should add an execution event to the action', async (test) => {
 	const actionCard = await test.context.jellyfish.getCardBySlug(
 		test.context.context, test.context.session, 'action-create-card')
 
-	const request = await test.context.worker.enqueue(test.context.session, {
+	const request = await test.context.queue.enqueue(test.context.session, {
 		action: actionCard.slug,
 		context: test.context.context,
 		card: typeCard.id,
@@ -490,7 +276,7 @@ ava('.execute() should execute a triggered action', async (test) => {
 		}
 	])
 
-	const request = await test.context.worker.enqueue(test.context.session, {
+	const request = await test.context.queue.enqueue(test.context.session, {
 		action: actionCard.slug,
 		context: test.context.context,
 		card: typeCard.id,
@@ -591,7 +377,7 @@ ava('.execute() should not execute a triggered action with a future start date',
 		}
 	])
 
-	const request = await test.context.worker.enqueue(test.context.session, {
+	const request = await test.context.queue.enqueue(test.context.session, {
 		action: actionCard.slug,
 		context: test.context.context,
 		card: typeCard.id,
@@ -664,7 +450,7 @@ ava('.execute() should execute a triggered action with a top level anyOf', async
 		}
 	])
 
-	const request = await test.context.worker.enqueue(test.context.session, {
+	const request = await test.context.queue.enqueue(test.context.session, {
 		action: actionCard.slug,
 		context: test.context.context,
 		card: typeCard.id,
@@ -695,7 +481,7 @@ ava('.execute() should add a create event when creating a card', async (test) =>
 	const actionCard = await test.context.jellyfish.getCardBySlug(
 		test.context.context, test.context.session, 'action-create-card')
 
-	const request = await test.context.worker.enqueue(test.context.session, {
+	const request = await test.context.queue.enqueue(test.context.session, {
 		action: actionCard.slug,
 		context: test.context.context,
 		card: typeCard.id,
@@ -742,7 +528,7 @@ ava('.execute() should add a create event when creating a card', async (test) =>
 ava('.execute() should be able to AGGREGATE based on the card timeline', async (test) => {
 	const typeType = await test.context.jellyfish.getCardBySlug(test.context.context, test.context.session, 'type')
 
-	const request = await test.context.worker.enqueue(test.context.session, {
+	const request = await test.context.queue.enqueue(test.context.session, {
 		action: 'action-create-card',
 		context: test.context.context,
 		card: typeType.id,
@@ -783,7 +569,7 @@ ava('.execute() should be able to AGGREGATE based on the card timeline', async (
 		test.context.context, request)
 	test.false(typeResult.error)
 
-	const threadRequest = await test.context.worker.enqueue(test.context.session, {
+	const threadRequest = await test.context.queue.enqueue(test.context.session, {
 		action: 'action-create-card',
 		context: test.context.context,
 		card: typeResult.data.id,
@@ -804,7 +590,7 @@ ava('.execute() should be able to AGGREGATE based on the card timeline', async (
 		test.context.context, threadRequest)
 	test.false(threadResult.error)
 
-	const messageRequest1 = await test.context.worker.enqueue(test.context.session, {
+	const messageRequest1 = await test.context.queue.enqueue(test.context.session, {
 		action: 'action-create-event',
 		context: test.context.context,
 		card: threadResult.data.id,
@@ -817,7 +603,7 @@ ava('.execute() should be able to AGGREGATE based on the card timeline', async (
 		}
 	})
 
-	const messageRequest2 = await test.context.worker.enqueue(test.context.session, {
+	const messageRequest2 = await test.context.queue.enqueue(test.context.session, {
 		action: 'action-create-event',
 		context: test.context.context,
 		card: threadResult.data.id,
@@ -845,7 +631,7 @@ ava('.execute() should be able to AGGREGATE based on the card timeline', async (
 ava('.execute() AGGREGATE should create a property on the target if it does not exist', async (test) => {
 	const typeType = await test.context.jellyfish.getCardBySlug(test.context.context, test.context.session, 'type')
 
-	const request = await test.context.worker.enqueue(test.context.session, {
+	const request = await test.context.queue.enqueue(test.context.session, {
 		action: 'action-create-card',
 		context: test.context.context,
 		card: typeType.id,
@@ -886,7 +672,7 @@ ava('.execute() AGGREGATE should create a property on the target if it does not 
 		test.context.context, request)
 	test.false(typeResult.error)
 
-	const threadRequest = await test.context.worker.enqueue(test.context.session, {
+	const threadRequest = await test.context.queue.enqueue(test.context.session, {
 		action: 'action-create-card',
 		context: test.context.context,
 		card: typeResult.data.id,
@@ -905,7 +691,7 @@ ava('.execute() AGGREGATE should create a property on the target if it does not 
 		test.context.context, threadRequest)
 	test.false(threadResult.error)
 
-	const messageRequest = await test.context.worker.enqueue(test.context.session, {
+	const messageRequest = await test.context.queue.enqueue(test.context.session, {
 		action: 'action-create-event',
 		context: test.context.context,
 		card: threadResult.data.id,
@@ -931,7 +717,7 @@ ava('.execute() AGGREGATE should create a property on the target if it does not 
 ava('.execute() AGGREGATE should work with $$ prefixed properties', async (test) => {
 	const typeType = await test.context.jellyfish.getCardBySlug(test.context.context, test.context.session, 'type')
 
-	const request = await test.context.worker.enqueue(test.context.session, {
+	const request = await test.context.queue.enqueue(test.context.session, {
 		action: 'action-create-card',
 		context: test.context.context,
 		card: typeType.id,
@@ -972,7 +758,7 @@ ava('.execute() AGGREGATE should work with $$ prefixed properties', async (test)
 		test.context.context, request)
 	test.false(typeResult.error)
 
-	const threadRequest = await test.context.worker.enqueue(test.context.session, {
+	const threadRequest = await test.context.queue.enqueue(test.context.session, {
 		action: 'action-create-card',
 		context: test.context.context,
 		card: typeResult.data.id,
@@ -993,7 +779,7 @@ ava('.execute() AGGREGATE should work with $$ prefixed properties', async (test)
 		test.context.context, threadRequest)
 	test.false(threadResult.error)
 
-	const messageRequest = await test.context.worker.enqueue(test.context.session, {
+	const messageRequest = await test.context.queue.enqueue(test.context.session, {
 		action: 'action-create-event',
 		card: threadResult.data.id,
 		context: test.context.context,
@@ -1020,7 +806,7 @@ ava('.execute() AGGREGATE should work with $$ prefixed properties', async (test)
 ava('.execute() should create a message with tags', async (test) => {
 	const typeType = await test.context.jellyfish.getCardBySlug(test.context.context, test.context.session, 'type')
 
-	const request = await test.context.worker.enqueue(test.context.session, {
+	const request = await test.context.queue.enqueue(test.context.session, {
 		action: 'action-create-card',
 		context: test.context.context,
 		card: typeType.id,
@@ -1051,7 +837,7 @@ ava('.execute() should create a message with tags', async (test) => {
 		test.context.context, request)
 	test.false(typeResult.error)
 
-	const threadRequest = await test.context.worker.enqueue(test.context.session, {
+	const threadRequest = await test.context.queue.enqueue(test.context.session, {
 		action: 'action-create-card',
 		context: test.context.context,
 		card: typeResult.data.id,
@@ -1069,7 +855,7 @@ ava('.execute() should create a message with tags', async (test) => {
 		test.context.context, threadRequest)
 	test.false(threadResult.error)
 
-	const messageRequest = await test.context.worker.enqueue(test.context.session, {
+	const messageRequest = await test.context.queue.enqueue(test.context.session, {
 		action: 'action-create-event',
 		context: test.context.context,
 		card: threadResult.data.id,
@@ -1798,7 +1584,7 @@ ava('.tick() should enqueue two actions if there are two time triggers with a pa
 
 ava('should be able to login as a user with a password', async (test) => {
 	const typeCard = await test.context.jellyfish.getCardBySlug(test.context.context, test.context.session, 'user')
-	const createUserRequest = await test.context.worker.enqueue(test.context.session, {
+	const createUserRequest = await test.context.queue.enqueue(test.context.session, {
 		action: 'action-create-user',
 		card: typeCard.id,
 		context: test.context.context,
@@ -1818,7 +1604,7 @@ ava('should be able to login as a user with a password', async (test) => {
 		test.context.context, createUserRequest)
 	test.false(signupResult.error)
 
-	const loginRequest = await test.context.worker.enqueue(test.context.session, {
+	const loginRequest = await test.context.queue.enqueue(test.context.session, {
 		action: 'action-create-session',
 		card: signupResult.data.id,
 		context: test.context.context,
@@ -1867,7 +1653,7 @@ ava('should be able to login as a password-less user', async (test) => {
 		}
 	})
 
-	const loginRequest = await test.context.worker.enqueue(test.context.session, {
+	const loginRequest = await test.context.queue.enqueue(test.context.session, {
 		action: 'action-create-session',
 		context: test.context.context,
 		card: user.id,
@@ -1900,7 +1686,7 @@ ava('should not be able to login as a password-less disallowed user', async (tes
 		}
 	})
 
-	await test.context.worker.enqueue(test.context.session, {
+	await test.context.queue.enqueue(test.context.session, {
 		action: 'action-create-session',
 		context: test.context.context,
 		card: user.id,
@@ -1917,7 +1703,7 @@ ava('should not be able to login as a password-less disallowed user', async (tes
 
 ava('should fail if signing up with the wrong password', async (test) => {
 	const typeCard = await test.context.jellyfish.getCardBySlug(test.context.context, test.context.session, 'user')
-	const createUserRequest = await test.context.worker.enqueue(test.context.session, {
+	const createUserRequest = await test.context.queue.enqueue(test.context.session, {
 		action: 'action-create-user',
 		context: test.context.context,
 		card: typeCard.id,
@@ -1937,7 +1723,7 @@ ava('should fail if signing up with the wrong password', async (test) => {
 		test.context.context, createUserRequest)
 	test.false(signupResult.error)
 
-	await test.context.worker.enqueue(test.context.session, {
+	await test.context.queue.enqueue(test.context.session, {
 		action: 'action-create-session',
 		context: test.context.context,
 		card: signupResult.data.id,
@@ -1959,7 +1745,7 @@ ava('should fail if signing up with the wrong password', async (test) => {
 
 ava('should fail to update a card if the schema does not match', async (test) => {
 	const typeCard = await test.context.jellyfish.getCardBySlug(test.context.context, test.context.session, 'card')
-	const request = await test.context.worker.enqueue(test.context.session, {
+	const request = await test.context.queue.enqueue(test.context.session, {
 		action: 'action-create-card',
 		context: test.context.context,
 		card: typeCard.id,
@@ -1980,7 +1766,7 @@ ava('should fail to update a card if the schema does not match', async (test) =>
 		test.context.context, request)
 	test.false(result.error)
 
-	await test.context.worker.enqueue(test.context.session, {
+	await test.context.queue.enqueue(test.context.session, {
 		action: 'action-update-card',
 		context: test.context.context,
 		card: result.data.id,
@@ -2000,7 +1786,7 @@ ava('should fail to update a card if the schema does not match', async (test) =>
 
 ava('should update a card to add an extra property', async (test) => {
 	const typeCard = await test.context.jellyfish.getCardBySlug(test.context.context, test.context.session, 'card')
-	const createRequest = await test.context.worker.enqueue(test.context.session, {
+	const createRequest = await test.context.queue.enqueue(test.context.session, {
 		action: 'action-create-card',
 		context: test.context.context,
 		card: typeCard.id,
@@ -2021,7 +1807,7 @@ ava('should update a card to add an extra property', async (test) => {
 		test.context.context, createRequest)
 	test.false(createResult.error)
 
-	const updateRequest = await test.context.worker.enqueue(test.context.session, {
+	const updateRequest = await test.context.queue.enqueue(test.context.session, {
 		action: 'action-update-card',
 		context: test.context.context,
 		card: createResult.data.id,
@@ -2058,7 +1844,7 @@ ava('should update a card to add an extra property', async (test) => {
 
 ava('should update a card to set active to false', async (test) => {
 	const typeCard = await test.context.jellyfish.getCardBySlug(test.context.context, test.context.session, 'card')
-	const createRequest = await test.context.worker.enqueue(test.context.session, {
+	const createRequest = await test.context.queue.enqueue(test.context.session, {
 		action: 'action-create-card',
 		context: test.context.context,
 		card: typeCard.id,
@@ -2076,7 +1862,7 @@ ava('should update a card to set active to false', async (test) => {
 		test.context.context, createRequest)
 	test.false(createResult.error)
 
-	const updateRequest = await test.context.worker.enqueue(test.context.session, {
+	const updateRequest = await test.context.queue.enqueue(test.context.session, {
 		action: 'action-update-card',
 		context: test.context.context,
 		card: createResult.data.id,
@@ -2108,7 +1894,7 @@ ava('should update a card to set active to false', async (test) => {
 
 ava('should update a card to set active to false using the card slug as input', async (test) => {
 	const typeCard = await test.context.jellyfish.getCardBySlug(test.context.context, test.context.session, 'card')
-	const createRequest = await test.context.worker.enqueue(test.context.session, {
+	const createRequest = await test.context.queue.enqueue(test.context.session, {
 		action: 'action-create-card',
 		context: test.context.context,
 		card: typeCard.id,
@@ -2126,7 +1912,7 @@ ava('should update a card to set active to false using the card slug as input', 
 		test.context.context, createRequest)
 	test.false(createResult.error)
 
-	const updateRequest = await test.context.worker.enqueue(test.context.session, {
+	const updateRequest = await test.context.queue.enqueue(test.context.session, {
 		action: 'action-update-card',
 		context: test.context.context,
 		card: 'foo-bar-baz',
@@ -2158,7 +1944,7 @@ ava('should update a card to set active to false using the card slug as input', 
 
 ava('should update a card to override an array property', async (test) => {
 	const typeCard = await test.context.jellyfish.getCardBySlug(test.context.context, test.context.session, 'card')
-	const createRequest = await test.context.worker.enqueue(test.context.session, {
+	const createRequest = await test.context.queue.enqueue(test.context.session, {
 		action: 'action-create-card',
 		context: test.context.context,
 		card: typeCard.id,
@@ -2179,7 +1965,7 @@ ava('should update a card to override an array property', async (test) => {
 		test.context.context, createRequest)
 	test.false(createResult.error)
 
-	const updateRequest = await test.context.worker.enqueue(test.context.session, {
+	const updateRequest = await test.context.queue.enqueue(test.context.session, {
 		action: 'action-update-card',
 		context: test.context.context,
 		card: createResult.data.id,
@@ -2216,7 +2002,7 @@ ava('should update a card to override an array property', async (test) => {
 
 ava('should add an update event if updating a card', async (test) => {
 	const typeCard = await test.context.jellyfish.getCardBySlug(test.context.context, test.context.session, 'card')
-	const createRequest = await test.context.worker.enqueue(test.context.session, {
+	const createRequest = await test.context.queue.enqueue(test.context.session, {
 		action: 'action-create-card',
 		context: test.context.context,
 		card: typeCard.id,
@@ -2237,7 +2023,7 @@ ava('should add an update event if updating a card', async (test) => {
 		test.context.context, createRequest)
 	test.false(createResult.error)
 
-	const updateRequest = await test.context.worker.enqueue(test.context.session, {
+	const updateRequest = await test.context.queue.enqueue(test.context.session, {
 		action: 'action-update-card',
 		context: test.context.context,
 		card: createResult.data.id,
@@ -2349,7 +2135,7 @@ ava('should add an update event if updating a card', async (test) => {
 
 ava('should delete a card using action-delete-card', async (test) => {
 	const typeCard = await test.context.jellyfish.getCardBySlug(test.context.context, test.context.session, 'card')
-	const createRequest = await test.context.worker.enqueue(test.context.session, {
+	const createRequest = await test.context.queue.enqueue(test.context.session, {
 		action: 'action-create-card',
 		context: test.context.context,
 		card: typeCard.id,
@@ -2367,7 +2153,7 @@ ava('should delete a card using action-delete-card', async (test) => {
 		test.context.context, createRequest)
 	test.false(createResult.error)
 
-	const deleteRequest = await test.context.worker.enqueue(test.context.session, {
+	const deleteRequest = await test.context.queue.enqueue(test.context.session, {
 		action: 'action-delete-card',
 		context: test.context.context,
 		card: createResult.data.id,
@@ -2394,7 +2180,7 @@ ava('should delete a card using action-delete-card', async (test) => {
 
 ava('should delete a card using action-update-card', async (test) => {
 	const typeCard = await test.context.jellyfish.getCardBySlug(test.context.context, test.context.session, 'card')
-	const createRequest = await test.context.worker.enqueue(test.context.session, {
+	const createRequest = await test.context.queue.enqueue(test.context.session, {
 		action: 'action-create-card',
 		context: test.context.context,
 		card: typeCard.id,
@@ -2412,7 +2198,7 @@ ava('should delete a card using action-update-card', async (test) => {
 		test.context.context, createRequest)
 	test.false(createResult.error)
 
-	const updateRequest = await test.context.worker.enqueue(test.context.session, {
+	const updateRequest = await test.context.queue.enqueue(test.context.session, {
 		action: 'action-update-card',
 		context: test.context.context,
 		card: createResult.data.id,
@@ -2445,7 +2231,7 @@ ava('should delete a card using action-update-card', async (test) => {
 ava('should post an error execute event if logging in as a disallowed user', async (test) => {
 	const adminCard = await test.context.jellyfish.getCardBySlug(test.context.context, test.context.session, 'user-admin')
 
-	const loginRequest = await test.context.worker.enqueue(test.context.session, {
+	const loginRequest = await test.context.queue.enqueue(test.context.session, {
 		action: 'action-create-session',
 		context: test.context.context,
 		card: adminCard.id,
@@ -2481,7 +2267,7 @@ ava('should post an error execute event if logging in as a disallowed user', asy
 ava('action-create-event should create a link card', async (test) => {
 	const typeCard = await test.context.jellyfish.getCardBySlug(test.context.context, test.context.session, 'card')
 
-	const cardRequest = await test.context.worker.enqueue(test.context.session, {
+	const cardRequest = await test.context.queue.enqueue(test.context.session, {
 		action: 'action-create-card',
 		context: test.context.context,
 		card: typeCard.id,
@@ -2496,7 +2282,7 @@ ava('action-create-event should create a link card', async (test) => {
 		test.context.context, cardRequest)
 	test.false(cardResult.error)
 
-	const messageRequest = await test.context.worker.enqueue(test.context.session, {
+	const messageRequest = await test.context.queue.enqueue(test.context.session, {
 		action: 'action-create-event',
 		context: test.context.context,
 		card: cardResult.data.id,
@@ -2575,7 +2361,7 @@ ava('events should always inherit their parent\'s markers', async (test) => {
 		type: 'type'
 	})
 
-	const cardRequest = await test.context.worker.enqueue(session, {
+	const cardRequest = await test.context.queue.enqueue(session, {
 		action: 'action-create-card',
 		context: test.context.context,
 		card: typeCard.id,
@@ -2592,7 +2378,7 @@ ava('events should always inherit their parent\'s markers', async (test) => {
 		test.context.context, cardRequest)
 	test.false(cardResult.error)
 
-	const messageRequest = await test.context.worker.enqueue(session, {
+	const messageRequest = await test.context.queue.enqueue(session, {
 		action: 'action-create-event',
 		context: test.context.context,
 		card: cardResult.data.id,
@@ -2625,7 +2411,7 @@ ava('Updating a cards markers should update the markers of attached events', asy
 		type: 'type'
 	})
 
-	const cardRequest = await test.context.worker.enqueue(session, {
+	const cardRequest = await test.context.queue.enqueue(session, {
 		action: 'action-create-card',
 		context: test.context.context,
 		card: typeCard.id,
@@ -2640,7 +2426,7 @@ ava('Updating a cards markers should update the markers of attached events', asy
 		test.context.context, cardRequest)
 	test.false(cardResult.error)
 
-	const messageRequest = await test.context.worker.enqueue(session, {
+	const messageRequest = await test.context.queue.enqueue(session, {
 		action: 'action-create-event',
 		context: test.context.context,
 		card: cardResult.data.id,
@@ -2659,7 +2445,7 @@ ava('Updating a cards markers should update the markers of attached events', asy
 		test.context.context, messageRequest)
 	test.false(messageResult.error)
 
-	const updateRequest = await test.context.worker.enqueue(session, {
+	const updateRequest = await test.context.queue.enqueue(session, {
 		action: 'action-update-card',
 		context: test.context.context,
 		card: cardResult.data.id,
@@ -2718,7 +2504,7 @@ ava('should be able to insert a deeply nested card', async (test) => {
 	}
 
 	const typeCard = await test.context.jellyfish.getCardBySlug(test.context.context, test.context.session, 'card')
-	const createRequest = await test.context.worker.enqueue(test.context.session, {
+	const createRequest = await test.context.queue.enqueue(test.context.session, {
 		action: 'action-create-card',
 		context: test.context.context,
 		card: typeCard.id,
@@ -2779,7 +2565,7 @@ ava('should be able to upsert a deeply nested card', async (test) => {
 	}
 
 	const typeCard = await test.context.jellyfish.getCardBySlug(test.context.context, test.context.session, 'card')
-	const createRequest = await test.context.worker.enqueue(test.context.session, {
+	const createRequest = await test.context.queue.enqueue(test.context.session, {
 		action: 'action-create-card',
 		context: test.context.context,
 		card: typeCard.id,
@@ -2798,7 +2584,7 @@ ava('should be able to upsert a deeply nested card', async (test) => {
 		test.context.context, createRequest)
 	test.false(createResult.error)
 
-	const updateRequest = await test.context.worker.enqueue(test.context.session, {
+	const updateRequest = await test.context.queue.enqueue(test.context.session, {
 		action: 'action-update-card',
 		context: test.context.context,
 		card: createResult.data.id,
